Cleanup and move rgb->yuv conversions into yuv file
