+++ src/lib.rs
@@ -24,10 +24,10 @@ mod lossless;
 mod lossless_transform;
 mod transform;
 mod vp8_arithmetic_decoder;
-mod yuv;
 mod vp8_arithmetic_encoder;
 mod vp8_common;
 mod vp8_encoder;
 mod vp8_prediction;
+mod yuv;
 
 pub mod vp8;
